magento/adobe-stock-integration#1157: Fixed static tests
