+++ MediaGallerySynchronization/composer.json
@@ -4,6 +4,7 @@
     "require": {
         "php": "~7.1.3||~7.2.0||~7.3.0",
         "magento/framework": "*",
+        "magento/module-media-gallery-api": "*",
         "magento/module-media-gallery-synchronization-api": "*"
     },
     "type": "magento2-module",
